requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
pandas==2.2.2
plotly==5.22.0
gunicorn==22.0.0
python-dotenv=1.0.1
